reorg: pull bilu into separate source files
