+++ CMakeLists_files.cmake
@@ -159,6 +159,7 @@ list (APPEND MAIN_SOURCE_FILES
   opm/simulators/linalg/FlowLinearSolverParameters.cpp
   opm/simulators/linalg/mixed/HaugenLabs.c
   opm/simulators/linalg/mixed/bsr.c
+  opm/simulators/linalg/mixed/prec.c
   opm/simulators/linalg/ISTLSolver.cpp
   opm/simulators/linalg/MILU.cpp
   opm/simulators/linalg/ParallelIstlInformation.cpp
